Support custom generator (#46)
* argocd: guard against missing Git configuration

If the found ApplicationSet contains no Git configuration Telefonistka
crashes.

Guarding against the missing configuration should avoid the crash.

* Make test parallel

* Use component path that works in lab

* Support custom generator

Adds support for custom generator as long as the generator has an input
parameter named "path".

Note that it would probably be advisable to make it configurable, but
this is left as a future exercise for the reader.